debugging numba implementation

+++ implementations/gpu_pathfinder.py
@@ -33,6 +33,6 @@ def GPUPathfinder(grid, start, goal, hArray):
     if x < grid.shape[0] and y < grid.shape[1]:
         goal_x, goal_y = goal
         if grid[x, y] != 0:
-            hArray[x, y] = heuristic(np.asarray((x,y)), goal)
+            hArray[x, y] = heuristic(np.asarray((x,y), dtype=np.int32), goal)
